+++ bittytax/conv/parsers/__init__.py
@@ -3,6 +3,8 @@ from . import barclays
 from . import bitfinex
 from . import bitstamp
 from . import bittrex
+from . import blockfi
+from . import celsius
 from . import cgtcalculator
 from . import changetip
 from . import circle