+++ CHANGELOG.md
@@ -42,6 +42,8 @@ Important:- A new Note field has been added to the end of the transaction record
 - Conversion tool: added dictionary to DataRow.
 - Conversion tool: "Savings & Loans" parser category added.
 - Conversion tool: convert_currency method added to DataParser.
+- Conversion tool: added parser for BlockFi.
+- Conversion tool: added parser for Celsius.
 ### Changed
 - Conversion tool: UnknownAddressError exception changed to generic DataFilenameError.
 - Binance parser: use filename to determine if deposits or withdrawals.